fixed dtos
