+++ src/main/java/faang/school/projectservice/dto/internship/InternshipFilterDto.java
@@ -1,7 +1,7 @@
-package faang.school.projectservice.dto.client.internship;
+package faang.school.projectservice.dto.internship;
 
+import faang.school.projectservice.model.InternshipStatus;
 import faang.school.projectservice.model.TeamRole;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,11 +13,8 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-public class InternshipUpdateDto {
+public class InternshipFilterDto {
 
-    @NotNull
-    private Long internshipId;
-
-    @NotNull
-    private TeamRole internNewTeamRole;
+    private InternshipStatus internshipStatus;
+    private TeamRole teamRole;
 }
