+++ src/main/java/faang/school/projectservice/dto/internship/InternshipUpdateDto.java
@@ -1,7 +1,7 @@
-package faang.school.projectservice.dto.client.internship;
+package faang.school.projectservice.dto.internship;
 
-import faang.school.projectservice.model.InternshipStatus;
 import faang.school.projectservice.model.TeamRole;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,10 +15,11 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-public class InternshipUpdateRequestDto {
+public class InternshipUpdateDto {
 
-    private Long id;
-    private List<Long> idsOfUsersWithCompletedTasks;
+    @NotNull
+    private Long internshipId;
+
+    @NotNull
     private TeamRole internNewTeamRole;
-    private InternshipStatus internshipStatus;
 }
