+++ src/main/java/faang/school/projectservice/dto/internship/InternshipUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.dto.client.internship;
+package faang.school.projectservice.dto.internship;
 
 import faang.school.projectservice.model.InternshipStatus;
 import faang.school.projectservice.model.TeamRole;
@@ -8,13 +8,18 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-public class InternshipFilterDto {
+public class InternshipUpdateRequestDto {
 
+    private Long id;
+    private List<Long> completedInternUserIds;
+    private List<Long> sackedInternUserIds;
+    private TeamRole internNewTeamRole;
     private InternshipStatus internshipStatus;
-    private TeamRole teamRole;
 }
