+++ src/main/java/faang/school/projectservice/filter/internship/InternshipTeamRoleFilter.java
@@ -1,6 +1,6 @@
 package faang.school.projectservice.filter.internship;
 
-import faang.school.projectservice.dto.client.internship.InternshipFilterDto;
+import faang.school.projectservice.dto.internship.InternshipFilterDto;
 import faang.school.projectservice.model.Internship;
 
 import java.util.stream.Stream;
